Renamed /effects -> /hooks and split the old ones between viewport.js and tracking.js
