+++ web/src/components/countrytable.js
@@ -12,11 +12,11 @@ import { precisionPrefix, formatPrefix } from 'd3-format';
 import { isArray, isFinite, noop } from 'lodash';
 
 import { dispatch, dispatchApplication } from '../store';
-import { useWidthObserver } from '../effects';
+import { useWidthObserver } from '../hooks/viewport';
 import {
   useCurrentZoneData,
   useCurrentZoneExchangeKeys,
-} from '../effects/redux';
+} from '../hooks/redux';
 import { getCo2Scale } from '../helpers/scales';
 import { getTooltipPosition } from '../helpers/graph';
 import { modeOrder, modeColor, DEFAULT_FLAG_SIZE } from '../helpers/constants';
