+++ web/src/components/countrytableoverlayifnodata.js
@@ -3,7 +3,7 @@ import { connect, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 
 import { __ } from '../helpers/translation';
-import { useCurrentZoneData } from '../effects/redux';
+import { useCurrentZoneData } from '../hooks/redux';
 
 const mapStateToProps = state => ({
   zoneTimeIndex: state.application.selectedZoneTimeIndex,
