+++ web/src/components/horizontalcolorbar.js
@@ -4,7 +4,7 @@ import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
 
 import { getCo2Scale } from '../helpers/scales';
-import { useWidthObserver, useHeightObserver } from '../effects';
+import { useWidthObserver, useHeightObserver } from '../hooks/viewport';
 
 const PADDING_X = 13;
 const PADDING_Y = 10;
