+++ web/src/hooks/viewport.js
@@ -1,7 +1,4 @@
 import { useState, useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-
-import thirdPartyServices from '../services/thirdparty';
 
 export function useWidthObserver(ref, offset = 0) {
   const [width, setWidth] = useState(0);
@@ -50,14 +47,3 @@ export function useHeightObserver(ref, offset = 0) {
 
   return height;
 }
-
-export const usePageViewsTracker = () => {
-  const { pathname, search } = useLocation();
-
-  // Update GA config whenever the URL changes.
-  useEffect(() => {
-    if (thirdPartyServices._ga) {
-      thirdPartyServices._ga.config({ page_path: `${pathname}${search}` });
-    }
-  }, [pathname, search]);
-};
